Extract out activity thread pool for general use
